Update cpp example to use vector<double>, correct CMakeLists.txt

+++ examples/c++/CIG.cpp
@@ -36,7 +36,7 @@ int main(){
         GameStatePtr state = game->getState();
         // Analyze the state.
 
-        std::vector<int> action(game->getAvailableButtonsSize());
+        std::vector<double> action(game->getAvailableButtonsSize());
         // Set your action.
 
         game->makeAction(action);
